backend/src/requirements.txt
appnope==0.1.3
asgiref==3.5.0
asttokens==2.0.5
async-timeout==4.0.2
backcall==0.2.0
decorator==5.1.1
Deprecated==1.2.13
Django==4.0.3
django-cors-headers==3.11.0
django-crispy-forms==1.14.0
django-environ==0.8.1
django-extensions==3.1.5
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
executing==0.8.3
gunicorn==20.1.0
importlib-metadata==4.11.3
ipdb==0.13.9
ipython==8.2.0
jedi==0.18.1
Markdown==3.3.6
matplotlib-inline==0.1.3
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
prompt-toolkit==3.0.29
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.11.2
PyJWT==2.4.0
pyparsing==3.0.9
python-environ==0.4.54
pytz==2022.1
PyYAML==6.0
redis==4.3.1
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
toml==0.10.2
traitlets==5.1.1
uritemplate==4.1.1
wcwidth==0.2.5
wrapt==1.14.1
zipp==3.8.0
